+++ assets/js/likes.js
@@ -56,3 +56,4 @@ $(document).ready(function() // HTMLが全て読み込まれたあと実行
     });
   });
 });
+
