--- design/page2.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
-        <title>jQuery PageMe demo</title>
-        <link rel="shortcut icon" href="favicon.ico">
-        <link rel="icon" href="favicon.ico">
-        <link rel="stylesheet" href="../assets/css/screen.css" type="text/css" media="screen">
-    </head>
-    <body>
-        <div class="header">
-            <a class="logo" href="index.html" title="PageMe">
-                <img src="../assets/images/pageme.png" alt="PageMe">
-            </a>
-        </div>
-        <div class="content">
-            <div class="container">
-                <h1>Image page demo</h1>
-                <img src="../assets/images/demo.png" alt="demo image">
-                <img src="../assets/images/demo.png" alt="demo image">
-                <img src="../assets/images/demo.png" alt="demo image">
-                <img src="../assets/images/demo.png" alt="demo image">
-                <img src="../assets/images/demo.png" alt="demo image">
-                <img src="../assets/images/demo.png" alt="demo image">
-            </div>
-            <div class="navigate">
-                <a class="prev" href="page1.html">Previous</a>
-            </div>
-        </div>
-        <div class="footer">
-            <p><a href="http://pageme.atecmedia.com" title="PageMe page">PageMe</a> | Programmed by Jens Eldering | &copy; 2013 by <a href="http://www.atecmedia.com" title="Visit Atec Media">Atec Media</a> | <a href="https://github.com/Scribilicious/jquery.pageme" title="Visit GitHub project page">GitHub</a> | <a href="http://creativecommons.org/licenses/by-sa/3.0/" title="Read the license">License</a></p>
-        </div>
-    </body>
-</html>
