--- testjs.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-  <meta charset="utf-8">
-  <title></title>
-</head>
-<body>
-  あああああああ
-
-<script type="text/javascript">
-  var sum = 0;
-  var a = [1,2];
-  var b = [3,4];
-  var array = [a, b];
-  array.forEach(function(value) {
-    console.log(value['0']);
-  });
-</script>
-</body>
-</html>
